requirements.txt
fastapi==0.73.0
motor==2.5.1
pydantic[email]
python-decouple==3.6
uvicorn==0.17.4
python-dateutil==2.8.2
qiskit==0.36.1
python-jose==3.3.0
passlib==1.7.4
python-multipart==0.0.5
bcrypt==3.2.2
strangeworks==0.2.13
strangeworks-braket==0.1.3